refactor(app): split markdown text primitives per platform
Replace selectable-text shim + runtime Platform.OS check with three
platform files exporting MarkdownTextSpan and MarkdownParagraphView.
Metro picks the right file at build time; no runtime branching in the
markdown renderer.

Android paragraphs stay a <View>, not <Text>. RN Android's text engine
does accept inline View children (TextInlineViewPlaceholderSpan in
ReactBaseTextShadowNode), but it inlines them as one-character
placeholders which collapses block-level paragraph images. <View>
preserves the original layout; cross-inline selection within a
paragraph remains iOS-only because it requires a single UITextView
ancestor with no Android equivalent.
