+++ packages/app/src/components/markdown-text.d.ts
@@ -1,0 +1,1 @@
+export * from "./markdown-text.ios";
