+++ packages/app/src/components/message.tsx
@@ -4,14 +4,13 @@ import {
   Image,
   Pressable,
   ActivityIndicator,
-  Platform,
   type GestureResponderEvent,
   type LayoutChangeEvent,
   StyleProp,
   ViewStyle,
   type TextStyle,
 } from "react-native";
-import { UITextView } from "@/components/selectable-text";
+import { MarkdownParagraphView, MarkdownTextSpan } from "@/components/markdown-text";
 import * as React from "react";
 import {
   useState,
@@ -1529,11 +1528,7 @@ function MarkdownInheritedText({
     () => [inheritedStyles, textStyle, overrideStyle],
     [inheritedStyles, textStyle, overrideStyle],
   );
-  return (
-    <UITextView uiTextView selectable style={style}>
-      {children}
-    </UITextView>
-  );
+  return <MarkdownTextSpan style={style}>{children}</MarkdownTextSpan>;
 }
 
 interface MarkdownListItemContentProps {
@@ -1545,28 +1540,6 @@ const MARKDOWN_LIST_ITEM_CONTENT_FLEX: ViewStyle = { flex: 1, flexShrink: 1, min
 
 function MarkdownListItemContent({ contentStyle, children }: MarkdownListItemContentProps) {
   const style = useMemo(() => [contentStyle, MARKDOWN_LIST_ITEM_CONTENT_FLEX], [contentStyle]);
-  return <View style={style}>{children}</View>;
-}
-
-interface MarkdownParagraphViewProps {
-  paragraphStyle: ViewStyle;
-  children: ReactNode;
-}
-
-const MARKDOWN_PARAGRAPH_RESET: ViewStyle = { marginBottom: 0 };
-
-function MarkdownParagraphView({ paragraphStyle, children }: MarkdownParagraphViewProps) {
-  const style = useMemo(() => [paragraphStyle, MARKDOWN_PARAGRAPH_RESET], [paragraphStyle]);
-  if (Platform.OS === "ios") {
-    // ViewStyle is structurally compatible with TextStyle for the layout
-    // properties paragraphs use (margin, padding, alignment). Cast lets
-    // the same style flow into UITextView, which expects TextStyle.
-    return (
-      <UITextView uiTextView selectable style={style as StyleProp<TextStyle>}>
-        {children}
-      </UITextView>
-    );
-  }
   return <View style={style}>{children}</View>;
 }
