--- packages/app/src/components/selectable-text.d.ts
@@ -1,1 +1,0 @@
-export * from "./selectable-text.native";
